[GDPR NOTICE] extract metrics and analytics to mobx mode
