+++ src/App.js
@@ -19,18 +19,8 @@
 import React from 'react';
 import { Provider, observer } from 'mobx-react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { YMInitializer } from 'react-yandex-metrika';
 
 import AppRouter from './routes';
-
-const YM_CONFIG = {
-  defer: false,
-  clickmap: true,
-  trackLinks: true,
-  // accurateTrackBounce: true,
-  // webvisor: true,
-  trackHash: false
-};
 
 @observer
 class App extends React.Component {
@@ -38,13 +28,7 @@ class App extends React.Component {
     return (
       <Provider store={this.props.store}>
         <Router>
-          <YMInitializer
-            accounts={[50501221]}
-            options={YM_CONFIG}
-            version='2'
-          >
-            <AppRouter />
-          </YMInitializer>
+          <AppRouter />
         </Router>
       </Provider>
     );
