+++ src/routes.js
@@ -24,7 +24,7 @@ import {
 } from 'react-router-dom';
 
 import LoadingLogo from './containers/LoadingLogo';
-import PrivacyNotice from './containers/PrivacyNotice/PrivacyNotice';
+import AnalyticsWrapper from './containers/Analytics/AnalyticsWrapper';
 
 const Narrative = lazy(() => import('./pages/Narrative'));
 const NotFound = lazy(() => import('./pages/404'));
@@ -34,7 +34,7 @@ const Demo = lazy(() => import('./pages/Demo'));
 
 const AppRouter = () => (
   <Suspense fallback={<LoadingLogo />}>
-    <PrivacyNotice />
+    <AnalyticsWrapper />
     <Switch>
       <Route exact path='/' render={() => <Redirect to='/world' />} />
       <Route path='/404' render={() => <NotFound />} />
